requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
anthropic>=0.28.0
openai>=1.30.0

# ── PDF extraction ────────────────────────────────────────────────────────────
PyMuPDF>=1.24.0
pdfminer.six>=20221105

# ── DOCX extraction ───────────────────────────────────────────────────────────
python-docx>=1.1.0

# ── PDF generation ────────────────────────────────────────────────────────────
weasyprint>=60.0          # Primary: HTML → PDF (best quality)
xhtml2pdf>=0.2.11         # Fallback PDF renderer

# ── Utilities ────────────────────────────────────────────────────────────────
python-dateutil>=2.9.0
